trying out log in back function

+++ controllers/c_posts.php
@@ -62,7 +62,7 @@ class posts_controller extends base_controller {
             			ON posts.user_id = users_users.user_id_followed
         			INNER JOIN users 
             			ON posts.user_id = users.user_id
-        			WHERE users_users.user_id = '.$this->user->user_id;' or posts.user_id = '.$this->user->user_id;
+        			WHERE users_users.user_id = '.$this->user->user_id;
                          
                 $posts = DB::instance(DB_NAME)->select_rows($q);
                 
@@ -140,48 +140,73 @@ class posts_controller extends base_controller {
 		------------------------------------------------------------------------------------------*/
 		
      	public function delete($post_id) {
-        
-       			DB::instance(DB_NAME)->delete('posts','WHERE post_id ='.$post_id);
-       
-      			# Send them back to the homepage
-       			Router::redirect('/posts');
-    	}        
-    
+                $q= 'SELECT
+                        *
+                        FROM posts
+                        WHERE post_id = '.$post_id;
+                $post = DB::instance(DB_NAME)->select_row($q);
+                $poster_id = $post['user_id'];
+                if ($this->user->user_id == $poster_id) {
+                        #delete the post when post id match is found
+                        DB::instance(DB_NAME)->delete('posts','WHERE post_id ='.$post_id);
+                        
+                        #Then send user back to view posts
+                        Router::redirect('/posts');
+                }
+                else {
+                        echo 'no permission';
+                }
+        } 
 	/*---------------------------------------------------------------------------------------------
 	edit post view
 	*/
 
-     public function edit($post_id) {
-        # Set up view
-                $this->template->content = View::instance("v_posts_edit");
+     public function edit($post_id = 0) {
+                 # Set up view
+         			$this->template->content = View::instance("v_posts_edit");
+         		 #Check to see if the post id exists
+                 if($post_id < 1) {
+                	die('Post not found. Please go back to view <a href="/posts">here.</a>');
+                 }
                 
-                # Set up query to get all users
-                $q = 'SELECT * FROM posts where post_id = '.$post_id;
-                        
-                # Run query
-                $post = DB::instance(DB_NAME)->select_row($q);
-                
-                # Pass data to the view
-                $this->template->content->post = $post;
-                
-                # Render view
-                echo $this->template;
+                 else {
+                 # Set up query to get all users
+         		$q = 'SELECT * FROM posts WHERE post_id = '.$post_id;
+                                
+         		# Run query
+         		$post = DB::instance(DB_NAME)->select_row($q);
 
-    }        
-    
+         		# Pass data to the view
+         		$this->template->content->post = $post;
+                                
+         		# Render view
+         		echo $this->template;
+                }
+    } 
 	/*---------------------------------------------------------------------------------------------
 	edit post view
 	*/
     public function p_edit($post_id) {
-                
-                $content = $_POST['content'];
-                                
-                # Update their row in the DB with the new token
-        		$data = Array(
-                'content' => $content);
-        
-                DB::instance(DB_NAME)->update('posts',$data, 'WHERE post_id ='.$post_id);                
-                Router::redirect('/posts/');
-                
-        }                
+                #Finds the post with matching Post ID
+                $q= 'SELECT
+                        *
+                        FROM posts
+                        WHERE post_id = '.$post_id;
+                        
+                $post = DB::instance(DB_NAME)->select_row($q);
+                $poster_id = $post['user_id'];
+                # If the user id matches with the user who made the post, then allow editing
+                if ($this->user->user_id == $poster_id) {
+                        $content = $_POST['content'];
+                        # Update their row in the DB with the new token
+                        $data = Array(
+                                'content' => $content
+                        );
+                        DB::instance(DB_NAME)->update('posts',$data, 'WHERE post_id ='.$post_id);
+                        Router::redirect('/posts');
+                }
+                else {
+                        die('No Permission to edit. Please login <a href="/users/login">here.</a>');
+                }
+        }             
 }# eoc
